# docs/swagger.yaml
definitions:
...
  models.Branch:
    properties:
      id:
        type: string
      location:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
  models.CloseJournalEntryInput:
    properties:
      cash_left:
        type: integer
      terminal_income:
        type: integer
    type: object
...
  models.Error:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  models.InitiatorType:
    enum:
      - salary
      - rent
      - utilities
      - other
      - sale
      - supplier
      - bnpl
    type: string
    x-enum-comments:
      InitiatorTypeBNPL: buy now pay later BNPL transactions
    x-enum-varnames:
      - InitiatorTypeSalary
      - InitiatorTypeRent
      - InitiatorTypeUtilities
      - InitiatorTypeOther
      - InitiatorTypeSales
      - InitiatorTypeSupplier
      - InitiatorTypeBNPL
  models.Journal:
    properties:
      branch:
        $ref: "#/definitions/models.Branch"
      cash_left:
        type: integer
      date:
        type: string
      id:
        type: string
      operations:
        items:
          $ref: "#/definitions/models.Transaction"
        type: array
      shift_is_closed:
        type: boolean
      terminal_income:
        type: integer
      total:
        type: integer
    type: object
...
  models.NewJournalEntryInput:
    properties:
      branch_name_or_id:
        type: string
      date:
        type: string
    type: object
...
  models.PaymentMethod:
    enum:
      - cash
      - bank
      - terminal
      - online_payment
      - cheque
      - online_transfer
      - undefined
    type: string
    x-enum-varnames:
      - PaymentMethodCash
      - PaymentMethodBank
      - PaymentMethodTerminal
      - OnlineMobileAppPayment
      - Cheque
      - OnlineTransfer
      - PaymentMethodUndefined
...
  models.Transaction:
    properties:
      amount:
        type: integer
      branch_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      payment_method:
        $ref: "#/definitions/models.PaymentMethod"
      type:
        $ref: "#/definitions/models.InitiatorType"
      updated_at:
        type: string
    type: object
...
info:
  contact:
    email: hamidovaslon13@gmail.com
    name: API Support
    url: https://github.com/aslon1213/g4h_pos_erp
  description: This is a ERP/POS API for Magazin.
  title: Magazin ERP/POS API
  version: "1.0"
paths:
...
  /api/journals:
    post:
      consumes:
        - application/json
      description: Create a new journal entry for a branch
      parameters:
        - description: New Journal Entry Input
          in: body
          name: input
          required: true
          schema:
            $ref: "#/definitions/models.NewJournalEntryInput"
      produces:
        - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: "#/definitions/models.Journal"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/models.Error"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/models.Error"
      security:
        - BearerAuth: []
      summary: Create a new journal entry
      tags:
        - journals
  /api/journals/{journal_id}:
    get:
      consumes:
        - application/json
      description: Get a journal entry by its ID
      parameters:
        - description: Journal ID
          in: path
          name: journal_id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/models.Journal"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/models.Error"
      security:
        - BearerAuth: []
      summary: Get a journal entry by ID
      tags:
        - journals
  /api/journals/{journal_id}/close:
    post:
      consumes:
        - application/json
      description: Close a journal entry by updating its transactions
      parameters:
        - description: Journal ID
          in: path
          name: journal_id
          required: true
          type: string
        - description: Close Journal Entry Input
          in: body
          name: input
          required: true
          schema:
            $ref: "#/definitions/models.CloseJournalEntryInput"
      produces:
        - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: "#/definitions/models.Transaction"
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/models.Error"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/models.Error"
      security:
        - BearerAuth: []
      summary: Close a journal entry
      tags:
        - journals
...
  /api/journals/{journal_id}/reopen:
    post:
      consumes:
        - application/json
      description: Reopen a journal entry by removing its closing transactions
      parameters:
        - description: Journal ID
          in: path
          name: journal_id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: "#/definitions/models.Transaction"
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/models.Error"
      security:
        - BearerAuth: []
      summary: Reopen a closed journal entry
      tags:
        - journals
  /api/journals/branch/{branch_id}:
    get:
      consumes:
        - application/json
      description: Query journal entries by branch ID
      parameters:
        - description: Branch ID
          in: path
          name: branch_id
          required: true
          type: string
        - description: Page number
          in: query
          name: page
          type: integer
        - description: Page size
          in: query
          name: page_size
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: "#/definitions/models.Journal"
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/models.Error"
      security:
        - BearerAuth: []
      summary: Query journal entries
      tags:
        - journals
  /api/journals/report:
    get:
      consumes:
        - application/json
      description: Get a report of journal entries
      produces:
        - application/json
      responses: {}
      security:
        - BearerAuth: []
      summary: Get a report
      tags:
        - journals
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"